server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.30.0
pytest==7.4.3
httpx==0.25.2
